new folder for trigonometric bloqs
