+++ docs/bloqs/index.rst
@@ -72,7 +72,7 @@ Bloqs Library
     arithmetic/conversions/contiguous_index.ipynb
     arithmetic/permutation.ipynb
     arithmetic/bitwise.ipynb
-    arithmetic/trigonometric.ipynb
+    arithmetic/trigonometric/trigonometric.ipynb
 
 .. toctree::
     :maxdepth: 2
