+++ qualtran/bloqs/arithmetic/trigonometric/arcsin_test.py
@@ -1,0 +1,28 @@
+#  Copyright 2024 Google LLC
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      https://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+
+import numpy as np
+
+from qualtran.bloqs.arithmetic.trigonometric.arcsin import _arcsin
+
+
+def test_arcsin(bloq_autotester):
+    bloq_autotester(_arcsin)
+
+
+def test_arcsin_classical():
+    bloq = _arcsin()
+    x = int(1 / np.sqrt(2) * 2**10)
+    result_out = int(np.arcsin(1 / np.sqrt(2)) * 2**10)
+    assert bloq.call_classically(x=x, result=0) == (x, result_out)
